requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
dataclasses-json==0.6.2
debugpy==1.8.0
decorator==5.1.1
distro==1.8.0
environs==9.5.0
exceptiongroup==1.2.0
executing==2.0.1
fonttools==4.47.2
frozenlist==1.4.0
greenlet==3.0.1
grpcio==1.58.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib-metadata==7.0.1
ipykernel==6.28.0
ipython==8.20.0
jedi==0.19.1
jiter==0.5.0
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
label-studio-sdk==0.0.32
label-studio-tools==0.0.3
langchain==0.1.4
langchain-community==0.0.16
langchain-core==0.2.36
langchain-openai==0.1.23
langmem==0.0.1rc4
langsmith==0.1.106
lxml==5.1.0
markdown-it-py==3.0.0
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
minio==7.2.0
mlcm==0.0.1
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.8
nltk==3.9.1
numpy==1.26.2
openai==1.42.0
orjson==3.10.7
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.0
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.6.2
pydantic_core==2.16.3
Pygments==2.17.2
pymilvus==2.3.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.2
rapidfuzz==3.7.0
regex==2023.10.3
requests==2.31.0
rich==13.7.0
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.2
tenacity==8.2.3
threadpoolctl==3.4.0
tiktoken==0.7.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.8.0
urllib3==2.1.0
wcwidth==0.2.13
wikipedia==1.4.0
yarl==1.9.2
zipp==3.17.0
